Expose a new on-disk format to improve the perf of docker_*
This change introduces two new utilities:
 - save.fast
 - docker_image.FromDisk

These duals roundtrip a more-efficient on-disk format for pulling and pushing images.  This also changes the puller/pusher to utilize them.  This is a breaking change to the signature of (at least) the docker_puller binary.

These changes are in support of a better intermediate format for Bazel's rules_docker, which already utilizes these tools to pull/build/push images without a Docker daemon.